Update components and stories
